move bookmark dialog to separate file
